+++ subtree/unit_test/include/boost/beast/unit_test/dstream.hpp
@@ -18,8 +18,8 @@
 #include <string>
 
 #ifdef BOOST_WINDOWS
-#include <boost/detail/winapi/basic_types.hpp>
-#include <boost/detail/winapi/debugapi.hpp>
+#include <boost/winapi/basic_types.hpp>
+#include <boost/winapi/debugapi.hpp>
 #endif
 
 namespace boost {
@@ -45,14 +45,14 @@ class dstream_buf
     void write(char const* s)
     {
         if(dbg_)
-            boost::detail::winapi::OutputDebugStringA(s);
+            boost::winapi::OutputDebugStringA(s);
         os_ << s;
     }
 
     void write(wchar_t const* s)
     {
         if(dbg_)
-            boost::detail::winapi::OutputDebugStringW(s);
+            boost::winapi::OutputDebugStringW(s);
         os_ << s;
     }
 
@@ -60,7 +60,7 @@ public:
     explicit
     dstream_buf(ostream& os)
         : os_(os)
-        , dbg_(boost::detail::winapi::IsDebuggerPresent() != 0)
+        , dbg_(boost::winapi::IsDebuggerPresent() != 0)
     {
     }